work around include issue with Magick++
On Debian bookworm I see this error with the current code:

In file included from /usr/include/boost/assert.hpp:58,
                 from /usr/include/boost/thread/pthread/mutex.hpp:22,
                 from /usr/include/boost/thread/mutex.hpp:16,
                 from include/ros/publisher.h:36,
                 from include/ros/node_handle.h:32,
                 from include/ros/ros.h:45,
                 from sr_tools/sr_movements/src/movement_from_image.cpp:31:
/usr/include/boost/thread/detail/platform_time.hpp: In static member function ‘static boost::detail::real_platform_timepoint boost::detail::real_platform_clock
::now()’:
/usr/include/boost/thread/detail/platform_time.hpp:311:11: error: ‘__assert_fail’ was not declared in this scope; did you mean ‘MagickCore::__assert_fail’?
  311 |           BOOST_ASSERT(0 && "Boost::Thread - clock_gettime(CLOCK_REALTIME) Internal Error");
      |           ^~~~~~~~~~~~

which is due to (Debian's?) Magick++ 6 including the headers in a namespace by accident.
While this is clearly a bug upstream as includes should be agnostic to order,
there is clearly no harm in changing the order here for a faster fix.